refactor: extract base class for LdapSocketFactory
- for our story (#182999703), the upgraded dependency removes
  [DummySSLSocketFactory](https://github.com/cloudfoundry/uaa/blob/f471118d103a157a5d9d76b60ad8caf944874ae5/server/src/main/java/org/cloudfoundry/identity/uaa/provider/ldap/ProcessLdapProperties.java#L40)
  class; hence we need to implement our own SSLSocketFactory that
  skips SSL validation.
- To do that, we created base class `BaseLdapSocketFactory` that `LdapSocketFactory`
  extends in this commit so that in the next commit we could create a
  new "SkipSslLdapSocketFactory" that extends `BaseLdapSocketFactory`.
  This approach helps us reuse shared code.

[#182999703]
