+++ server/src/main/java/org/cloudfoundry/identity/uaa/security/BaseLdapSocketFactory.java
@@ -1,0 +1,55 @@
+package org.cloudfoundry.identity.uaa.security;
+
+import javax.net.SocketFactory;
+import javax.net.ssl.SSLSocketFactory;
+import java.io.IOException;
+import java.net.InetAddress;
+import java.net.Socket;
+
+public class BaseLdapSocketFactory extends SSLSocketFactory {
+    private static SocketFactory instance;
+    protected SSLSocketFactory delegate;
+
+    public static SocketFactory getDefault() {
+        if (instance == null) {
+            instance = new LdapSocketFactory();
+        }
+
+        return instance;
+    }
+
+    @Override
+    public String[] getDefaultCipherSuites() {
+        return delegate.getDefaultCipherSuites();
+    }
+
+    @Override
+    public String[] getSupportedCipherSuites() {
+        return delegate.getSupportedCipherSuites();
+    }
+
+    @Override
+    public Socket createSocket(Socket socket, String s, int i, boolean b) throws IOException {
+        return delegate.createSocket(socket, s, i, b);
+    }
+
+    @Override
+    public Socket createSocket(String s, int i) throws IOException {
+        return delegate.createSocket(s, i);
+    }
+
+    @Override
+    public Socket createSocket(String s, int i, InetAddress inetAddress, int i1) throws IOException {
+        return delegate.createSocket(s, i, inetAddress, i1);
+    }
+
+    @Override
+    public Socket createSocket(InetAddress inetAddress, int i) throws IOException {
+        return delegate.createSocket(inetAddress, i);
+    }
+
+    @Override
+    public Socket createSocket(InetAddress inetAddress, int i, InetAddress inetAddress1, int i1) throws IOException {
+        return delegate.createSocket(inetAddress, i, inetAddress1, i1);
+    }
+}
