+++ server/src/main/java/org/cloudfoundry/identity/uaa/security/LdapSocketFactory.java
@@ -23,18 +23,7 @@ import java.net.InetAddress;
 import java.net.Socket;
 import java.security.SecureRandom;
 
-public class LdapSocketFactory extends SSLSocketFactory {
-
-    private static SocketFactory instance;
-    private SSLSocketFactory delegate;
-
-    public static SocketFactory getDefault() {
-        if (instance == null) {
-            instance = new LdapSocketFactory();
-        }
-
-        return instance;
-    }
+public class LdapSocketFactory extends BaseLdapSocketFactory {
 
     public LdapSocketFactory() {
         try {
@@ -48,38 +37,4 @@ public class LdapSocketFactory extends SSLSocketFactory {
         }
     }
 
-    @Override
-    public String[] getDefaultCipherSuites() {
-        return delegate.getDefaultCipherSuites();
-    }
-
-    @Override
-    public String[] getSupportedCipherSuites() {
-        return delegate.getSupportedCipherSuites();
-    }
-
-    @Override
-    public Socket createSocket(Socket socket, String s, int i, boolean b) throws IOException {
-        return delegate.createSocket(socket, s, i, b);
-    }
-
-    @Override
-    public Socket createSocket(String s, int i) throws IOException {
-        return delegate.createSocket(s, i);
-    }
-
-    @Override
-    public Socket createSocket(String s, int i, InetAddress inetAddress, int i1) throws IOException {
-        return delegate.createSocket(s, i, inetAddress, i1);
-    }
-
-    @Override
-    public Socket createSocket(InetAddress inetAddress, int i) throws IOException {
-        return delegate.createSocket(inetAddress, i);
-    }
-
-    @Override
-    public Socket createSocket(InetAddress inetAddress, int i, InetAddress inetAddress1, int i1) throws IOException {
-        return delegate.createSocket(inetAddress, i, inetAddress1, i1);
-    }
 }
